requirements.txt
Flask==2.1.1
gevent==21.12.0
keras==2.8.0
numpy==1.22.3
Pillow==9.1.0
Werkzeug==2.1.1
